Move AWS placebo fixtures into test/units/utils so module_utils tests can use them
