+++ test/units/modules/cloud/amazon/test_aws_direct_connect_connection.py
@@ -15,7 +15,7 @@
 # You should have received a copy of the GNU General Public License
 # along with Ansible.  If not, see <http://www.gnu.org/licenses/>.
 
-from . placebo_fixtures import placeboify, maybe_sleep
+from units.utils.amazon_placebo_fixtures import placeboify, maybe_sleep
 from ansible.modules.cloud.amazon import aws_direct_connect_connection
 from ansible.module_utils.ec2 import get_aws_connection_info, boto3_conn
