+++ test/units/modules/cloud/amazon/test_aws_direct_connect_link_aggregation_group.py
@@ -18,7 +18,7 @@
 import pytest
 import os
 import collections
-from . placebo_fixtures import placeboify, maybe_sleep
+from units.utils.amazon_placebo_fixtures import placeboify, maybe_sleep
 from ansible.modules.cloud.amazon import aws_direct_connect_link_aggregation_group as lag_module
 from ansible.module_utils.ec2 import get_aws_connection_info, boto3_conn
